requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1607109633063/work
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
astor==0.8.1
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1606255099073/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
charset-normalizer==3.3.2
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
coloredlogs==15.0
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1607706933335/work
decorator==4.4.2
et-xmlfile==1.0.1
fonttools==4.38.0
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1597424467487/work
h5py==2.8.0
humanfriendly==9.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
jmespath==0.10.0
keras==2.11.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
Markdown @ file:///tmp/build/80754af9/markdown_1605111062842/work
matplotlib==3.5.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456400975/work
natsort @ file:///tmp/build/80754af9/natsort_1606153679261/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
opencv-python-headless==4.8.0.76
openpyxl==3.0.5
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==23.1
pandas==1.2.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.18
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.19.0
Pygments==2.8.1
PyJWT @ file:///tmp/build/80754af9/pyjwt_1608676183569/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==5.3.1
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-image==0.19.3
scikit-learn==0.20.3
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
six @ file:///tmp/build/80754af9/six_1605205313296/work
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorboardX==2.1
tensorflow==2.1.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor @ file:///croot/termcolor_1668084651543/work
tifffile==2021.11.2
toml==0.10.2
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.3.1
torchsummary==1.5.1
torchvision==0.4.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.65.2
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt @ file:///tmp/build/80754af9/wrapt_1638420625940/work
yarl @ file:///tmp/build/80754af9/yarl_1598045053219/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
